requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
instagrapi==2.1.2
moviepy==2.0.0
numpy==2.1.3
pillow==10.4.0
proglog==0.1.10
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
pysrt==1.1.2
python-dotenv==1.0.1
pytube==15.0.0
pytubefix==8.5.1
requests==2.32.3
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3